🚚 Split android module into device and package
Trying to keep separate things separate. Also, move the functions that check if ADB and AAPT2 are present into `main`.
